+++ Source/SoundManager.h
@@ -2,13 +2,16 @@
 #include "SDL_mixer.h"
 
 #include <unordered_map>
-
+#include <unordered_set>
 
 
 class SoundMan
 {
 	private:
 		inline static std::unordered_map<std::string, Mix_Chunk*> Sounds;
+
+		static void RefreshSoundsInFolder(const std::string& directory, bool removeInvalid, std::unordered_set<std::string>& namesCollector);
+
 	public:
 		static void Innit();
 
@@ -35,6 +38,10 @@ class SoundMan
 		static void SetVolume(const std::string& soundKey, unsigned char volume);
 
 		static bool DeleteSound(const std::string& name);
+
+		// Loads new previously unloaded sounds in slected folder and all recursive folders
+		// removeInvalid flag will remove sounds that no longer exist WARNING this will break exsting pointers
+		static void RefreshSounds(const std::string& directory, bool removeInvalid = false);
 
 		static void Clear();
 };